# Detecting and linking against libgmp