Hi,
I hope this can solve #118

I try to link against all the default paths I could think of, not sure how important is this but I wanted to try and make it as robust as possible.


Would love if people can try and run it on their machines and hopefully also machines *without* gmp to make sure it works correctly.


FYI I ran some benchmarks:
without gmp:
```
test benches::bench_sign       ... bench:      37,741 ns/iter (+/- 2,715)
test benches::bench_verify     ... bench:      56,191 ns/iter (+/- 1,943)
test benches::generate         ... bench:      22,192 ns/iter (+/- 1,482)
test ecdh::benches::bench_ecdh ... bench:      61,198 ns/iter (+/- 3,360)


test benches::bench_sign       ... bench:      37,081 ns/iter (+/- 4,044)
test benches::bench_verify     ... bench:      58,233 ns/iter (+/- 2,599)
test benches::generate         ... bench:      22,114 ns/iter (+/- 946)
test ecdh::benches::bench_ecdh ... bench:      61,475 ns/iter (+/- 3,203)
```

with gmp:
```
test benches::bench_sign       ... bench:      39,418 ns/iter (+/- 2,815)
test benches::bench_verify     ... bench:      49,183 ns/iter (+/- 1,411)
test benches::generate         ... bench:      22,517 ns/iter (+/- 390)
test ecdh::benches::bench_ecdh ... bench:      61,905 ns/iter (+/- 1,251)

test benches::bench_sign       ... bench:      38,450 ns/iter (+/- 3,045)
test benches::bench_verify     ... bench:      49,546 ns/iter (+/- 1,447)
test benches::generate         ... bench:      22,758 ns/iter (+/- 3,818)
test ecdh::benches::bench_ecdh ... bench:      61,198 ns/iter (+/- 1,918)

test benches::bench_sign       ... bench:      38,054 ns/iter (+/- 3,151)
test benches::bench_verify     ... bench:      48,028 ns/iter (+/- 738)
test benches::generate         ... bench:      22,130 ns/iter (+/- 901)
test ecdh::benches::bench_ecdh ... bench:      61,243 ns/iter (+/- 741)
```

looks like ~15% improvement in verification time.

cc @apoelstra