@elichai 2019-07-23
Yeah, I'll try to look into Windows build system and figure out how do I do that. (I've never compiled C/C++ on windows)