@apoelstra 2019-07-23
Are your benchmarks with or without endo on? On what hardware?