@elichai 2019-07-23
hmm without.
On my laptop, CPU: i9-9980HK, NVMe 1TB, 32GB RAM.
I'll retry with endo on