@apoelstra 2019-07-23
Oh, nice. No worries - I was just surprised because the numbers looked so good. But you have a very powerful laptop, that's all.