@elichai 2019-07-23
With both gmp and endo:
```
test benches::bench_sign       ... bench:      39,821 ns/iter (+/- 10,845)
test benches::bench_verify     ... bench:      33,353 ns/iter (+/- 4,957)
test benches::generate         ... bench:      22,466 ns/iter (+/- 1,123)
test ecdh::benches::bench_ecdh ... bench:      46,736 ns/iter (+/- 1,998)

test benches::bench_sign       ... bench:      38,214 ns/iter (+/- 3,916)
test benches::bench_verify     ... bench:      32,790 ns/iter (+/- 2,845)
test benches::generate         ... bench:      22,433 ns/iter (+/- 4,472)
test ecdh::benches::bench_ecdh ... bench:      46,293 ns/iter (+/- 3,522)

```