@real-or-random 2019-07-23
The endomorphism is certainly turned off because of patent concerns. For GMP I don't know the original reasoning but a good argument is that external libraries are very annoying when it comes to consensus critical code, e.g., OpenSSL has been a problem in the past. Imagine a new GMP version has a bug (or just behaves differently than old versions) and people start to link against it. This could end up badly. Of course a new secp256k1 version could have a bug as well but secp256k1 is developed with its very specific application requirements in mind whereas GMP is simply a general purpose library.