@apoelstra 2019-07-24
GMP is turned off because it's a pain in the ass to use external deps without autotools, that's all.