@elichai 2019-07-24
@apoelstra but bitcoin uses autotools