@elichai 2019-07-25
I launched a debian LXDE VM with no gmp, downloaded cargo from the debian repository and this is what I get:
http://paste.debian.net/plain/1093103

It compiles fine but print a warning about the failed compilation attempt(`check_gmp`).
I also added prints/panics to make sure it didn't detect gmp,
then I installed gmp and tried again and it detected it correctly.