@elichai 2019-07-25
Damn. I just realized that both this and `check_uint128_t` destroy cross compilation.
although the gmp situation is a bit worse.

If you compile on a system that have gmp but you then run the code on a system that don't have it's a problem.

build.rs is compiled using the host toolchain, not the target toolchain.

I'll look into https://doc.rust-lang.org/cargo/reference/environment-variables.html#environment-variables-cargo-sets-for-build-scripts and run this only when we're on one of these:
1. `i686-apple-darwin`, `i686-unknown-linux-gnu`, `x86_64-apple-darwin`, `x86_64-unknown-linux-gnu`.