@elichai 2019-07-25
DONT MERGE.
this breaks cross compilation right now.
i'll fix it tomorrow.