@elichai 2019-07-26
Ok,
So it seems that `cc` only builds static libraries, this in reality just checks for the include headers because AFAIK there's no linker involved in static libraries.

what I did try to do is to gate this check only for targets where this could work. 
that way when you compile to a target where this doesn't really makes sense it won't even try.

For example, before it would try to compile `check_gmp` when the target for wasm, it would succeed if the host have the right header files but it will fail on link time because you don't have gmp compiled for wasm on your machine.

(I retested it on wasm and now it doesn't try to link to gmp anymore)

this still isn't even close to perfect, because we basically just check that the headers exist on the host computer :/