@elichai 2019-09-05
Just realized something. this is not fun for cross compiling.
but I think this is also the case in the native autotools.
(a host might have libgmp but the target might not)