@elichai 2020-12-22
I don't think we care about this anymore now that https://github.com/bitcoin-core/secp256k1/pull/831 is getting close to merging